backend/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
sqlalchemy[asyncio]>=2.0.25
asyncpg>=0.29.0
alembic>=1.13.0
pydantic>=2.6.0
pydantic-settings>=2.1.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
httpx>=0.27.0
redis>=5.0.0
openai>=1.12.0
anthropic>=0.18.0
google-generativeai>=0.4.0
python-multipart>=0.0.9
pgvector>=0.2.5
groq>=0.13.0
tavily-python>=0.5.0
sse-starlette>=2.0.0
